+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.19.1
 
 * Improving `tokenizers/sentences/naive`.
-* Dropping `generatorics` in favoir of [obliterator](https://github.com/Yomguithereal/obliterator).
+* Dropping `generatorics` in favor of [obliterator](https://github.com/Yomguithereal/obliterator).
 
 ## 0.19.0